backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "httpx[http2]>=0.27.0",
    "langchain-community>=0.3.20",
    "langchain-core>=0.3.20",
    "langchain-openai>=0.3.5",
    "langchain-pinecone>=0.2.13",
    "langchain-text-splitters>=0.3.20",
    "openai>=1.0.0",
    "pinecone>=3.0.0",
    "pydantic>=2.0.0",
    "pypdf>=4.3.1",
    "python-dotenv==1.0.1",
    "regex>=2024.11.6",
    "tqdm>=4.67.1",
    "uvicorn[standard]>=0.32.0",
    "redis>=5.0.0",
    "langsmith[openai-agents]>=0.1.0",
    "rank-bm25>=0.2.2",
    "pdfplumber==0.11.9",
    "pytesseract==0.3.13",
    "pdf2image==1.17.0",
    "langchain>=1.2.15",
]
